+++ src/CosmosDB/CosmosDB/Models/Restore/PSBackupInformation.cs
@@ -14,15 +14,15 @@
 
 using Microsoft.Azure.Management.CosmosDB.Models;
 
-namespace Microsoft.Azure.PowerShell.Cmdlets.CosmosDB.Models.Restore.Sql
+namespace Microsoft.Azure.PowerShell.Cmdlets.CosmosDB.Models.Restore
 {
-    public class PSSqlBackupInformation
+    public class PSBackupInformation
     {
-        public PSSqlBackupInformation()
+        public PSBackupInformation()
         {
         }
 
-        public PSSqlBackupInformation(BackupInformation backupInformation)
+        public PSBackupInformation(BackupInformation backupInformation)
         {
             if (backupInformation == null || backupInformation.ContinuousBackupInformation == null)
                 return;
@@ -32,7 +32,7 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.CosmosDB.Models.Restore.Sql
 
         //
         // Summary:
-        //     Gets sql container latest restorable timestamp.
+        //     Gets latest restorable timestamp.
         public string LatestRestorableTimestamp { get; set; }
     }
 }